ref: Clean up configs
