+++ vite.main.config.js
@@ -1,6 +1,25 @@
 import { defineConfig, mergeConfig } from "vite";
 
-import { getBuildConfig, getBuildDefine, external, pluginHotRestart } from "./vite.base.config.js";
+import { getBuildConfig, getDefineKeys, external, pluginHotRestart } from "./vite.base.config.js";
+
+/** @type {(env: import('vite').ConfigEnv<'build'>) => Record<string, any>} */
+const getBuildDefine = (env) => {
+  const { command, forgeConfig } = env;
+
+  const names = forgeConfig.renderer.filter(({ name }) => name != null).map(({ name }) => name);
+  const defineKeys = getDefineKeys(names);
+  const define = Object.entries(defineKeys).reduce((acc, [name, keys]) => {
+    const { VITE_DEV_SERVER_URL, VITE_NAME } = keys;
+    const def = {
+      [VITE_DEV_SERVER_URL]:
+        command === "serve" ? JSON.stringify(process.env[VITE_DEV_SERVER_URL]) : undefined,
+      [VITE_NAME]: JSON.stringify(name),
+    };
+    return { ...acc, ...def };
+  }, {});
+
+  return define;
+};
 
 export default defineConfig((env) => {
   return mergeConfig(getBuildConfig(env), {
